+++ src/xyz/oihana/schema/constants/traits/business/DocumentsTrait.php
@@ -5,10 +5,12 @@ namespace xyz\oihana\schema\constants\traits\business;
 use xyz\oihana\schema\constants\traits\business\documents\AdjustmentTrait;
 use xyz\oihana\schema\constants\traits\business\documents\AppliedEcoFeeTrait;
 use xyz\oihana\schema\constants\traits\business\documents\BusinessDocumentLineTrait;
+use xyz\oihana\schema\constants\traits\business\documents\BusinessDocumentTrait;
 use xyz\oihana\schema\constants\traits\business\documents\DocumentTotalsTrait;
 use xyz\oihana\schema\constants\traits\business\documents\EcoFeeRuleTrait;
 use xyz\oihana\schema\constants\traits\business\documents\PaymentInstallmentTrait;
 use xyz\oihana\schema\constants\traits\business\documents\PaymentScheduleTrait;
+use xyz\oihana\schema\constants\traits\business\documents\QuoteTrait;
 use xyz\oihana\schema\constants\traits\business\documents\TaxDetailTrait;
 
 /**
@@ -23,9 +25,11 @@ trait DocumentsTrait
     use AdjustmentTrait          ,
         AppliedEcoFeeTrait       ,
         BusinessDocumentLineTrait ,
+        BusinessDocumentTrait    ,
         DocumentTotalsTrait      ,
         EcoFeeRuleTrait          ,
         PaymentInstallmentTrait  ,
         PaymentScheduleTrait     ,
+        QuoteTrait               ,
         TaxDetailTrait           ;
 }